Simplified main.py and Split Overrides
