+++ rules/examples/inheritance_rules.py
